LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/pseudo-on-scroller-expected.txt


FAIL scroll nearest on pseudo-element attaches to parent scroll container assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
